requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colored==2.2.4
contourpy==1.2.1
cycler==0.12.1
fonttools==4.52.1
gdpc==7.2.0
idna==3.7
imageio==2.34.1
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.0
more-itertools==10.2.0
NBT==1.5.1
nbtlib==2.0.4
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
PyGLM==2.7.1
pyglm-typing==0.2.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.2
scikit-image==0.23.2
scipy==1.13.1
six==1.16.0
termcolor==2.4.0
tifffile==2024.5.22
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
